AUTO_CLICK/requirements.txt
customtkinter
pyautogui
opencv-python
numpy
Pillow
mss
pynput
